simplify boxing/unboxing

+++ solr/core/src/test/org/apache/solr/cloud/OverseerCollectionConfigSetProcessorTest.java
@@ -922,7 +922,7 @@ public class OverseerCollectionConfigSetProcessorTest extends SolrTestCaseJ4 {
             + numberOfNodesToCreateOn
             + " is not allowed to be higher than numberOfNodes "
             + numberOfNodes,
-        numberOfNodes.intValue() >= numberOfNodesToCreateOn.intValue());
+            numberOfNodes >= numberOfNodesToCreateOn);
     assertTrue(
         "Wrong usage of testTemplate. createNodeListOption has to be "
             + CreateNodeListOptions.SEND
